Change data rep from Gitlab HTTP to SourceForce GIT
HTTP download is too unreliable for such a large download. To make matters worse Git is unable to resume an incomplete download.
GitLab doesnt offer cloning using Git protocol if you dont have an account on gitlab and keys set.
So I'm changing this to SouceForce's Git repo and Git protocol. Which is a lot more reliable.